Reorg of Scan
I realized that I was sending the issue data as unique columns and not a json package. It made my code wayyyy longer than it needed to be. pausing on the scan to get the BigQuery syncs set up. Once those are done, going to get the /up running, then scans, then crawl.